backend/requirements.txt
numpy==1.26.4
openai==0.28.0
python-dotenv==1.0.1
Flask==3.0.3
faiss-cpu==1.9.0
pdf2image==1.17.0
pandas==2.2.3
pytesseract==0.3.13
